requirements.txt
aiogram==3.18.0
aiosqlite==0.21.0
APScheduler==3.10.4
httpx==0.28.1
python-dotenv==1.0.1
